Generalised checkValidity, moved sample functions to UnivariateDistribution, fixed NIG pdf

+++ distributions/univariate/discrete/BernoulliRand.cpp
@@ -62,17 +62,17 @@ double BernoulliRand::Entropy()
     return -(p * std::log(p) + q * std::log(q));
 }
 
-bool BernoulliRand::fitMLE(const std::vector<double> &sample)
+bool BernoulliRand::fitMLE(const std::vector<int> &sample)
 {
     return fitProbabilityMLE(sample);
 }
 
-bool BernoulliRand::fitMM(const std::vector<double> &sample)
+bool BernoulliRand::fitMM(const std::vector<int> &sample)
 {
     return fitProbabilityMM(sample);
 }
 
-bool BernoulliRand::fitBayes(const std::vector<double> &sample, BetaRand &priorDistribution)
+bool BernoulliRand::fitBayes(const std::vector<int> &sample, BetaRand &priorDistribution)
 {
     return fitProbabilityBayes(sample, priorDistribution);
 }
